move ClientPoolImpl to core
