+++ core/src/main/java/org/apache/iceberg/ClientPoolImpl.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iceberg.hive;
+package org.apache.iceberg;
 
 import java.io.Closeable;
 import java.util.ArrayDeque;
@@ -37,7 +37,7 @@ public abstract class ClientPoolImpl<C, E extends Exception> implements Closeabl
   private volatile int currentSize;
   private boolean closed;
 
-  ClientPoolImpl(int poolSize, Class<? extends E> reconnectExc) {
+  public ClientPoolImpl(int poolSize, Class<? extends E> reconnectExc) {
     this.poolSize = poolSize;
     this.reconnectExc = reconnectExc;
     this.clients = new ArrayDeque<>(poolSize);
